убрал функцианал для доски из app
